@@ -1,12 +1,12 @@
 {
   "name": "signalk-server",
-  "version": "2.7.2",
+  "version": "2.12.0",
   "description": "An implementation of a [Signal K](http://signalk.org) server for boats.",
   "main": "index.js",
   "scripts": {
     "build": "tsc --build && npm run build:docs",
     "build:all": "npm run build:workspaces && npm run build && npm run build-declaration && npm run build:docs",
-    "build:docs": "USER_INFO=$(id -u):$(id -g) docker run --rm -v ${PWD}/docs:/book peaceiris/mdbook:v0.4.34 build",
+    "build:docs": "USER_INFO=$(id -u):$(id -g) docker run --rm -v ${PWD}/docs:/book peaceiris/mdbook:v0.4.40 build",
     "build:workspaces": "npm run build --workspaces --if-present",
     "build-declaration": "tsc --declaration",
     "watch": "tsc --build -w",
@@ -74,11 +74,11 @@
     "@signalk/course-provider": "^1.0.0",
     "@signalk/n2k-signalk": "^3.0.0",
     "@signalk/nmea0183-signalk": "^3.0.0",
-    "@signalk/resources-provider": "^1.0.0",
-    "@signalk/server-admin-ui": "2.7.x",
-    "@signalk/server-api": "2.3.x",
+    "@signalk/resources-provider": "^1.3.0",
+    "@signalk/server-admin-ui": "2.11.x",
+    "@signalk/server-api": "2.5.x",
     "@signalk/signalk-schema": "^1.7.1",
-    "@signalk/streams": "^4.1.0",
+    "@signalk/streams": "^4.3.0",
     "api-schema-builder": "^2.0.11",
     "baconjs": "^1.0.1",
     "bcryptjs": "^2.4.3",
@@ -133,7 +133,6 @@
     "@signalk/signalk-to-nmea0183": "^1.0.0",
     "@signalk/vesselpositions": "^1.0.0",
     "@signalk/udp-nmea-plugin": "^2.0.0",
-    "@signalk/zones": "^1.0.0",
     "signalk-to-nmea2000": "^2.16.0",
     "signalk-n2kais-to-nmea0183": "^1.3.1",
     "mdns": "^2.5.1",
@@ -150,15 +149,13 @@
     "@types/mocha": "^10.0.1",
     "@types/ncp": "^2.0.5",
     "@types/node-fetch": "^2.5.3",
-    "@types/rmfr": "^2.0.1",
     "@types/semver": "^7.1.0",
     "@types/serialport": "^8.0.2",
     "@types/split": "^1.0.0",
     "@types/swagger-ui-express": "^4.1.3",
     "@types/unzipper": "^0.10.5",
     "@types/uuid": "^8.3.1",
-    "@typescript-eslint/eslint-plugin": "^5.38.0",
-    "@typescript-eslint/parser": "^5.38.0",
+    "@typescript-eslint/eslint-plugin": "^8.16.0",
     "chai": "^4.0.0",
     "chai-json-equal": "0.0.1",
     "chai-things": "^0.2.0",
@@ -170,11 +167,10 @@
     "lint-staged": "^10.0.3",
     "mocha": "^10.2.0",
     "prettier": "^2.7.1",
-    "rimraf": "^3.0.2",
-    "rmfr": "^2.0.0",
+    "rimraf": "^6.0.1",
     "ts-node": "^10.9.1",
-    "typedoc": "^0.23.23",
-    "typescript": "^4.8.3"
+    "typedoc": "^0.26.11",
+    "typescript": "5.6.x"
   },
   "funding": "https://opencollective.com/signalk"
 }
