@@ -1,6 +1,6 @@
 {
   "name": "@signalk/server-admin-ui",
-  "version": "2.7.0",
+  "version": "2.11.0",
   "description": "Signal K server admin webapp",
   "author": "Scott Bender, Teppo Kurki",
   "contributors": [
@@ -27,6 +27,8 @@
     "buffer": "^6.0.3",
     "copy-webpack-plugin": "^11.0.0",
     "css-loader": "^5.0.0",
+    "eslint": "^8.57.0",
+    "eslint-plugin-react": "^7.34.1",
     "file-loader": "^6.1.1",
     "font-awesome": "^4.7.0",
     "html-webpack-plugin": "^5.0.0-alpha.6",
@@ -34,9 +36,10 @@
     "lodash.get": "^4.4.2",
     "lodash.remove": "^4.7.0",
     "lodash.set": "^4.3.2",
+    "lodash.uniq": "^4.5.0",
     "moment": "^2.29.1",
-    "node-sass": "^8.0.0",
     "prettier": "^2.3.2",
+    "react": "^16.13.1",
     "react-copy-to-clipboard": "^5.0.3",
     "react-dom": "^16.13.1",
     "react-html-parser": "^2.0.2",
@@ -45,24 +48,25 @@
     "react-redux": "^5.1.2",
     "react-router-dom": "^4.3.1",
     "react-select": "^3.1.0",
-    "react": "^16.13.1",
     "reactstrap": "^5.0.0",
     "reconnecting-websocket": "^4.4.0",
-    "redux-thunk": "2.3.0",
     "redux": "^3.7.2",
-    "sass-loader": "^13.2.0",
+    "redux-thunk": "2.3.0",
+    "sass": "^1.81.0",
+    "sass-loader": "^16.0.3",
     "simple-line-icons": "^2.5.5",
     "style-loader": "^2.0.0",
+    "webpack": "^5.96.1",
     "webpack-bundle-analyzer": "^3.9.0",
-    "webpack-cli": "^4.2.0",
-    "webpack": "^5.0.0"
+    "webpack-cli": "^4.2.0"
   },
   "scripts": {
     "prepublishOnly": "npm run clean && npm run build",
     "watch": "webpack --watch --mode development",
     "build": "webpack --mode=production",
     "format": "prettier --write src/",
-    "ci": "prettier --check src/",
+    "eslint": "eslint src/",
+    "ci": "prettier --check src/ && npm run eslint",
     "clean": "rimraf ./public",
     "bundle-analyzer": "webpack-bundle-analyzer --port 4200 public/stats.json"
   }
