@@ -1,26 +1,20 @@
 import React, { Component } from 'react'
 import { connect } from 'react-redux'
-import get from 'lodash.get'
 import JSONTree from 'react-json-tree'
 import {
-  Badge,
-  Button,
   Card,
   CardHeader,
   CardBody,
-  CardFooter,
-  InputGroup,
-  InputGroupAddon,
   Input,
   Form,
   Col,
   Label,
   FormGroup,
-  FormText,
   Table,
 } from 'reactstrap'
 import moment from 'moment'
 import { CopyToClipboard } from 'react-copy-to-clipboard'
+import Meta from './Meta'
 
 const timestampFormat = 'MM/DD HH:mm:ss'
 
@@ -74,7 +68,7 @@ class DataBrowser extends Component {
     this.handleMessage = this.handleMessage.bind(this)
     this.handleContextChange = this.handleContextChange.bind(this)
     this.handleSearch = this.handleSearch.bind(this)
-    this.handleMeta = this.handleMeta.bind(this)
+    this.toggleMeta = this.toggleMeta.bind(this)
   }
 
   handleMessage(msg) {
@@ -213,7 +207,7 @@ class DataBrowser extends Component {
     localStorage.setItem(searchStorageKey, event.target.value)
   }
 
-  handleMeta(event) {
+  toggleMeta(event) {
     this.setState({ ...this.state, includeMeta: event.target.checked })
     localStorage.setItem(metaStorageKey, event.target.checked)
   }
@@ -271,7 +265,7 @@ class DataBrowser extends Component {
                       id="Meta"
                       name="meta"
                       className="switch-input"
-                      onChange={this.handleMeta}
+                      onChange={this.toggleMeta}
                       checked={this.state.includeMeta}
                     />
                     <span
@@ -357,7 +351,6 @@ class DataBrowser extends Component {
                           const meta =
                             this.state.meta[this.state.context][data.path]
                           const units = meta && meta.units ? meta.units : ''
-                          const path = key.substring(0, key.lastIndexOf('.'))
 
                           return (
                             <tr key={key}>
@@ -403,29 +396,31 @@ class DataBrowser extends Component {
                       </tr>
                     </thead>
                     <tbody>
-                      {Object.keys(this.state.meta[this.state.context] || {})
+                      {Object.keys(this.state.data[this.state.context] || {})
                         .filter((key) => {
                           return (
                             !this.state.search ||
                             this.state.search.length === 0 ||
-                            key.indexOf(this.state.search) !== -1
+                            key
+                              .toLowerCase()
+                              .indexOf(this.state.search.toLowerCase()) !== -1
                           )
                         })
+                        .map(
+                          (key) => this.state.data[this.state.context][key].path
+                        )
+                        .filter((path, index, array) => {
+                          //filter dups
+                          return array.indexOf(path) === index
+                        })
                         .sort()
-                        .map((key) => {
-                          const meta = this.state.meta[this.state.context][key]
-                          const formatted = JSON.stringify(meta, null, 2)
-                          const path = key
+                        .map((path) => {
+                          const meta = this.state.meta[this.state.context][path]
                           return (
                             <tr key={path}>
                               <td>{path}</td>
                               <td>
-                                <pre
-                                  className="text-primary"
-                                  style={{ whiteSpace: 'pre-wrap' }}
-                                >
-                                  {formatted}
-                                </pre>
+                                <Meta meta={meta || {}} path={path} />
                               </td>
                             </tr>
                           )
