@@ -48,9 +48,13 @@ function N2KAnalyzer(options) {
   const that = this
   this.linereader.on('line', function (data) {
     try {
-      parsed = JSON.parse(data)
+      let parsed = JSON.parse(data)
+      if (parsed.version) {
+        console.log('Connected to analyzer v' + parsed.version)
+        return
+      }
       that.push(parsed)
-      options.app.emit(this.analyzerOutEvent, parsed)
+      options.app.emit(that.analyzerOutEvent, parsed)
     } catch (ex) {
       console.error(data)
       console.error(ex.stack)
