@@ -20,7 +20,7 @@
 
 const Parser0183 = require('@signalk/nmea0183-signalk')
 const N2kMapper = require('@signalk/n2k-signalk').N2kMapper
-const { putPath } = require('../put')
+const { putPath, deletePath } = require('../put')
 const {
   isN2KString,
   FromPgn,
@@ -29,9 +29,15 @@ const {
 
 const serverRoutesPrefix = '/skServer'
 
+let n2kOutAvailable = false
+
 module.exports = function (app) {
   const n2kMapper = new N2kMapper({ app }, app.propertyValues)
   const pgnParser = new FromPgn({}, app.propertyValues)
+
+  app.on('nmea2000OutAvailable', () => {
+    n2kOutAvailable = true
+  })
 
   const processors = {
     n2k: (msgs, sendToServer) => {
@@ -46,7 +52,7 @@ module.exports = function (app) {
           return n2kMapper.toDelta(n2k)
         }
       })
-      return { deltas, n2kJson: n2kJson }
+      return { deltas, n2kJson: n2kJson, n2kOutAvailable }
     },
     '0183': (msgs) => {
       const parser = new Parser0183({ app })
@@ -67,7 +73,7 @@ module.exports = function (app) {
 
         if (first.pgn) {
           type = 'n2k-json'
-        } else if (first.updates || first.put) {
+        } else if (first.updates || first.put || first.delete) {
           type = 'signalk'
         } else {
           return { error: 'unknown JSON format' }
@@ -97,9 +103,10 @@ module.exports = function (app) {
 
   app.post(`${serverRoutesPrefix}/inputTest`, (req, res) => {
     const sendToServer = req.body.sendToServer
+    const sendToN2K = req.body.sendToN2K
 
     if (
-      sendToServer &&
+      (sendToServer || sendToN2K) &&
       !app.securityStrategy.isDummy() &&
       !app.securityStrategy.allowConfigure(req)
     ) {
@@ -111,6 +118,13 @@ module.exports = function (app) {
 
     if (error) {
       res.status(400).json({ error: error })
+      return
+    }
+
+    if (sendToN2K && type != 'n2k-json' && type != 'n2k') {
+      res.status(400).json({
+        error: 'Please enter NMEA 2000 json format or Actisense format'
+      })
       return
     }
 
@@ -139,6 +153,26 @@ module.exports = function (app) {
                 )
               })
             )
+          } else if (msg.delete) {
+            puts.push(
+              new Promise((resolve) => {
+                setTimeout(() => {
+                  resolve('Timed out waiting for put result')
+                }, 5000)
+                deletePath(
+                  app,
+                  msg.context,
+                  msg.delete.path,
+                  req,
+                  msg.requestId,
+                  (reply) => {
+                    if (reply.state !== 'PENDING') {
+                      resolve(reply)
+                    }
+                  }
+                )
+              })
+            )
           } else {
             app.handleMessage('input-test', msg)
           }
@@ -151,6 +185,12 @@ module.exports = function (app) {
       } else {
         res.json({ deltas: msgs })
       }
+    } else if (sendToN2K) {
+      const event = type == 'n2k' ? 'nmea2000out' : 'nmea2000JsonOut'
+      msgs.forEach((msg) => {
+        app.emit(event, msg)
+      })
+      res.json({ deltas: [] })
     } else {
       try {
         const data = processors[type](msgs, sendToServer)
