@@ -17,7 +17,6 @@
 
 import { Ports } from '@signalk/server-api'
 import fs from 'fs'
-import { SerialPort } from 'serialport'
 
 export const listAllSerialPorts = (): Promise<Ports> => {
   return new Promise((resolve, reject) => {
@@ -36,9 +35,9 @@ export const listAllSerialPorts = (): Promise<Ports> => {
 
 function listSerialPorts() {
   try {
-    // eslint-disable-next-line @typescript-eslint/no-var-requires
-    return SerialPort.list()
-  } catch (err) {
+    // eslint-disable-next-line @typescript-eslint/no-var-requires, @typescript-eslint/no-require-imports
+    return require('serialport').SerialPort.list()
+  } catch (_err) {
     return Promise.resolve([])
   }
 }
